Crate aggregator

Aggregator example for Project Oak.

This shows how an Oak Node can aggregate data samples and report aggregated values if there are enough samples to hide individual contributors (enforces k-anonymity).

Clients invoke the module by providing data samples that contain a bucket ID and a Sparse Vector - a dictionary with integer keys.

Structs

AggregatorNode
Oak Node that collects and aggregates data. Data is collected in the aggregators map where keys are buckets and values are instances of a ThresholdAggregator. Option is used to keep note of the outdated buckets: once an Aggregator has sent its data to the Backend, it's replaced with None and all subsequent client requests corresponding to its bucket are discarded.

Functions

grpc_oak_main

oak_main